requirements.txt
python-dotenv==1.0.0
python-decouple==3.8
requests==2.31.0
Flask==2.3.3
gunicorn==21.2.0
schedule==1.2.0